+++ stac_fastapi/pgstac/app.py
@@ -20,17 +20,22 @@ from stac_fastapi.api.models import (
     create_post_request_model,
     create_request_model,
 )
-from stac_fastapi.api.openapi import update_openapi
 from stac_fastapi.extensions.core import (
+    CollectionSearchExtension,
+    CollectionSearchFilterExtension,
     FieldsExtension,
-    FilterExtension,
     FreeTextExtension,
+    ItemCollectionFilterExtension,
     OffsetPaginationExtension,
+    SearchFilterExtension,
     SortExtension,
     TokenPaginationExtension,
     TransactionExtension,
 )
-from stac_fastapi.extensions.core.collection_search import CollectionSearchExtension
+from stac_fastapi.extensions.core.fields import FieldsConformanceClasses
+from stac_fastapi.extensions.core.free_text import FreeTextConformanceClasses
+from stac_fastapi.extensions.core.query import QueryConformanceClasses
+from stac_fastapi.extensions.core.sort import SortConformanceClasses
 from stac_fastapi.extensions.third_party import BulkTransactionExtension
 from starlette.middleware import Middleware
 
@@ -59,23 +64,32 @@ search_extensions_map = {
     "query": QueryExtension(),
     "sort": SortExtension(),
     "fields": FieldsExtension(),
-    "filter": FilterExtension(client=FiltersClient()),
+    "filter": SearchFilterExtension(client=FiltersClient()),
     "pagination": TokenPaginationExtension(),
 }
 
 # collection_search extensions
 cs_extensions_map = {
-    "query": QueryExtension(),
-    "sort": SortExtension(),
-    "fields": FieldsExtension(),
-    "filter": FilterExtension(client=FiltersClient()),
-    "free_text": FreeTextExtension(),
+    "query": QueryExtension(conformance_classes=[QueryConformanceClasses.COLLECTIONS]),
+    "sort": SortExtension(conformance_classes=[SortConformanceClasses.COLLECTIONS]),
+    "fields": FieldsExtension(conformance_classes=[FieldsConformanceClasses.COLLECTIONS]),
+    "filter": CollectionSearchFilterExtension(client=FiltersClient()),
+    "free_text": FreeTextExtension(
+        conformance_classes=[FreeTextConformanceClasses.COLLECTIONS],
+    ),
     "pagination": OffsetPaginationExtension(),
 }
 
 # item_collection extensions
 itm_col_extensions_map = {
-    "filter": FilterExtension(client=FiltersClient()),
+    "query": QueryExtension(
+        conformance_classes=[QueryConformanceClasses.ITEMS],
+    ),
+    "sort": SortExtension(
+        conformance_classes=[SortConformanceClasses.ITEMS],
+    ),
+    "fields": FieldsExtension(conformance_classes=[FieldsConformanceClasses.ITEMS]),
+    "filter": ItemCollectionFilterExtension(client=FiltersClient()),
     "pagination": TokenPaginationExtension(),
 }
 
@@ -123,6 +137,7 @@ if itm_col_extensions:
         extensions=itm_col_extensions,
         request_type="GET",
     )
+    application_extensions.extend(itm_col_extensions)
 
 # /collections model
 collections_get_request_model = EmptyRequest
@@ -145,17 +160,17 @@ async def lifespan(app: FastAPI):
     await close_db_connection(app)
 
 
-fastapp = FastAPI(
-    openapi_url=settings.openapi_url,
-    docs_url=settings.docs_url,
-    redoc_url=None,
-    root_path=settings.root_path,
-    lifespan=lifespan,
-)
-
-
 api = StacApi(
-    app=update_openapi(fastapp),
+    app=FastAPI(
+        openapi_url=settings.openapi_url,
+        docs_url=settings.docs_url,
+        redoc_url=None,
+        root_path=settings.root_path,
+        title=settings.stac_fastapi_title,
+        version=settings.stac_fastapi_version,
+        description=settings.stac_fastapi_description,
+        lifespan=lifespan,
+    ),
     settings=settings,
     extensions=application_extensions,
     client=CoreCrudClient(pgstac_search_model=post_request_model),